refactor: split misc tests by feature
