+++ tests/misc.rs
@@ -26,59 +26,6 @@ use users::{get_current_gid, get_current_uid, get_group_by_gid, get_user_by_uid}
 mod common;
 use common::read_golden;
 
-#[cfg(unix)]
-struct Tmpfs(TempDir);
-
-#[cfg(unix)]
-impl Tmpfs {
-    fn new() -> Option<Self> {
-        if !cfg!(target_os = "linux") {
-            return None;
-        }
-        if !Uid::effective().is_root() {
-            return None;
-        }
-        let mount_exists = std::env::var_os("PATH").is_some_and(|paths| {
-            std::env::split_paths(&paths).any(|dir| dir.join("mount").is_file())
-        });
-        if !mount_exists {
-            return None;
-        }
-        if let Ok(fs) = std::fs::read_to_string("/proc/filesystems") {
-            if !fs.lines().any(|l| l.trim().ends_with("tmpfs")) {
-                return None;
-            }
-        } else {
-            return None;
-        }
-        let dir = tempdir().ok()?;
-        let status = std::process::Command::new("mount")
-            .args(["-t", "tmpfs", "tmpfs", dir.path().to_str().unwrap()])
-            .stdout(std::process::Stdio::null())
-            .stderr(std::process::Stdio::null())
-            .status()
-            .ok()?;
-        if status.success() {
-            Some(Tmpfs(dir))
-        } else {
-            None
-        }
-    }
-
-    fn path(&self) -> &std::path::Path {
-        self.0.path()
-    }
-}
-
-#[cfg(unix)]
-impl Drop for Tmpfs {
-    fn drop(&mut self) {
-        let _ = std::process::Command::new("umount")
-            .arg(self.0.path())
-            .status();
-    }
-}
-
 #[allow(clippy::vec_init_then_push)]
 #[allow(clippy::vec_init_then_push)]
 #[test]
@@ -381,63 +328,6 @@ fn local_sync_without_flag_succeeds() {
 }
 
 #[test]
-fn leading_dash_directory_requires_separator() {
-    let dir = tempdir().unwrap();
-    let src = dir.path().join("-src");
-    std::fs::create_dir_all(src.join("sub")).unwrap();
-    fs::write(src.join("sub/file.txt"), b"dash").unwrap();
-
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .current_dir(dir.path())
-        .args(["-src/", "dst"])
-        .assert()
-        .failure();
-
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .current_dir(dir.path())
-        .args(["--", "-src/", "dst"])
-        .assert()
-        .success();
-
-    assert_eq!(
-        fs::read(dir.path().join("dst/sub/file.txt")).unwrap(),
-        b"dash"
-    );
-}
-
-#[test]
-fn colon_in_path_triggers_remote_mode() {
-    let dir = tempdir().unwrap();
-    fs::create_dir_all(dir.path().join("dst")).unwrap();
-    fs::write(dir.path().join("with:colon"), b"colon").unwrap();
-
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .current_dir(dir.path())
-        .env("RSYNC_RSH", "false")
-        .args(["with:colon", "dst"])
-        .assert()
-        .failure();
-}
-
-#[test]
-fn relative_preserves_ancestors() {
-    let dir = tempdir().unwrap();
-    let src_root = dir.path().join("src");
-    std::fs::create_dir_all(src_root.join("a/b")).unwrap();
-    std::fs::write(src_root.join("a/b/file.txt"), b"hi").unwrap();
-    let mut cmd = Command::cargo_bin("oc-rsync").unwrap();
-    cmd.current_dir(dir.path());
-    cmd.args(["-R", "src/a/b/", "dst"]);
-    cmd.assert().success();
-
-    let out = std::fs::read(dir.path().join("dst/src/a/b/file.txt")).unwrap();
-    assert_eq!(out, b"hi");
-}
-
-#[test]
 fn resumes_from_partial_dir() {
     let dir = tempdir().unwrap();
     let src_dir = dir.path().join("src");
@@ -540,1435 +430,15 @@ fn resumes_from_partial_file_with_temp_dir() {
 }
 
 #[test]
-fn fails_when_temp_dir_is_file() {
-    let dir = tempdir().unwrap();
-    let src_dir = dir.path().join("src");
-    let dst_dir = dir.path().join("dst");
-    let tmp_file = dir.path().join("tmp");
-    std::fs::create_dir_all(&src_dir).unwrap();
-    std::fs::write(src_dir.join("a.txt"), b"hello").unwrap();
-    std::fs::create_dir_all(&dst_dir).unwrap();
-    std::fs::write(&tmp_file, b"not a dir").unwrap();
-
-    let mut cmd = Command::cargo_bin("oc-rsync").unwrap();
-    let src_arg = format!("{}/", src_dir.display());
-    cmd.args([
-        "--temp-dir",
-        tmp_file.to_str().unwrap(),
-        &src_arg,
-        dst_dir.to_str().unwrap(),
-    ]);
-    cmd.assert().failure();
-    assert!(!dst_dir.join("a.txt").exists());
-}
-
-#[test]
-fn temp_files_created_in_temp_dir() {
-    let dir = tempdir().unwrap();
-    let src_dir = dir.path().join("src");
-    let dst_dir = dir.path().join("dst");
-    let tmp_dir = dir.path().join("tmp");
-    std::fs::create_dir_all(&src_dir).unwrap();
-    std::fs::create_dir_all(&dst_dir).unwrap();
-    std::fs::create_dir_all(&tmp_dir).unwrap();
-    let data = vec![b'x'; 200_000];
-    std::fs::write(src_dir.join("a.txt"), &data).unwrap();
-
-    let src_arg = format!("{}/", src_dir.display());
-    let mut child = std::process::Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args([
-            "--partial",
-            "--bwlimit",
-            "10240",
-            "--temp-dir",
-            tmp_dir.to_str().unwrap(),
-            &src_arg,
-            dst_dir.to_str().unwrap(),
-        ])
-        .spawn()
-        .unwrap();
-
-    let mut found = false;
-    for _ in 0..50 {
-        let tmp_present = std::fs::read_dir(&tmp_dir)
-            .unwrap()
-            .filter_map(|e| {
-                let entry = e.ok()?;
-                let name = entry.file_name();
-                let name = name.to_string_lossy();
-                if name.starts_with(".a.txt.") {
-                    Some(entry.path())
-                } else {
-                    None
-                }
-            })
-            .next();
-        if tmp_present.is_some() {
-            found = true;
-            break;
-        }
-        thread::sleep(Duration::from_millis(100));
-    }
-    let _ = child.kill();
-    let _ = child.wait();
-    assert!(found, "temp file not created in temp dir during transfer");
-    assert!(
-        fs::read_dir(&tmp_dir).unwrap().all(|e| !e
-            .unwrap()
-            .file_name()
-            .to_string_lossy()
-            .starts_with(".oc-rsync-tmp.")),
-        "intermediate temp dir created",
-    );
-}
-
 #[test]
 #[cfg(unix)]
-fn temp_dir_cross_filesystem_temp_file_in_dest() {
-    let base = tempdir_in(".").unwrap();
-    let src_dir = base.path().join("src");
-    let dst_dir = base.path().join("dst");
-    fs::create_dir_all(&src_dir).unwrap();
-    fs::create_dir_all(&dst_dir).unwrap();
-    let data = vec![b'x'; 200_000];
-    fs::write(src_dir.join("a.txt"), &data).unwrap();
-
-    let tmp_dir = match Tmpfs::new() {
-        Some(t) => t,
-        None => {
-            eprintln!("skipping cross-filesystem temp-dir test; tmpfs unavailable");
-            return;
-        }
-    };
-
-    let dst_dev = fs::metadata(&dst_dir).unwrap().dev();
-    let tmp_dev = fs::metadata(tmp_dir.path()).unwrap().dev();
-    assert_ne!(dst_dev, tmp_dev, "devices match");
-
-    let src_arg = format!("{}/", src_dir.display());
-    let mut child = std::process::Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args([
-            "--partial",
-            "--bwlimit",
-            "10240",
-            "--temp-dir",
-            tmp_dir.path().to_str().unwrap(),
-            &src_arg,
-            dst_dir.to_str().unwrap(),
-        ])
-        .spawn()
-        .unwrap();
-
-    let mut found = false;
-    for _ in 0..50 {
-        let tmp_present = fs::read_dir(&dst_dir)
-            .unwrap()
-            .filter_map(|e| {
-                let entry = e.ok()?;
-                let name = entry.file_name();
-                let name = name.to_string_lossy();
-                if name.starts_with(".a.txt.") {
-                    Some(entry.path())
-                } else {
-                    None
-                }
-            })
-            .next();
-        if tmp_present.is_some() {
-            assert!(fs::read_dir(tmp_dir.path()).unwrap().next().is_none());
-            found = true;
-            break;
-        }
-        thread::sleep(Duration::from_millis(100));
-    }
-    let _ = child.kill();
-    let _ = child.wait();
-    assert!(
-        found,
-        "temp file not created in destination during transfer"
-    );
-    assert!(
-        fs::read_dir(&dst_dir).unwrap().all(|e| !e
-            .unwrap()
-            .file_name()
-            .to_string_lossy()
-            .starts_with(".oc-rsync-tmp.")),
-        "intermediate temp dir created",
-    );
-
-    let out = fs::read(dst_dir.join("a.txt")).unwrap();
-    assert_eq!(out.len(), data.len());
-}
-
-#[test]
-fn destination_is_replaced_atomically() {
-    let dir = tempdir().unwrap();
-    let src_dir = dir.path().join("src");
-    let dst_dir = dir.path().join("dst");
-    std::fs::create_dir_all(&src_dir).unwrap();
-    std::fs::create_dir_all(&dst_dir).unwrap();
-    let src_file = src_dir.join("a.txt");
-    let dst_file = dst_dir.join("a.txt");
-    std::fs::write(&src_file, vec![b'x'; 50_000]).unwrap();
-    std::fs::write(&dst_file, b"old").unwrap();
-
-    let src_arg = format!("{}/", src_dir.display());
-    let mut child = std::process::Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args(["--bwlimit", "20000", &src_arg, dst_dir.to_str().unwrap()])
-        .spawn()
-        .unwrap();
-
-    let mut found = false;
-    let mut tmp_file: Option<PathBuf> = None;
-    for _ in 0..50 {
-        let tmp_present = fs::read_dir(&dst_dir)
-            .unwrap()
-            .filter_map(|e| {
-                let entry = e.ok()?;
-                let name = entry.file_name();
-                let name = name.to_string_lossy();
-                if name.starts_with(".a.txt.") {
-                    Some(entry.path())
-                } else {
-                    None
-                }
-            })
-            .next();
-        if let Some(path) = tmp_present {
-            let out = std::fs::read(&dst_file).unwrap();
-            assert_eq!(out, b"old");
-            tmp_file = Some(path);
-            found = true;
-            break;
-        }
-        thread::sleep(Duration::from_millis(100));
-    }
-    assert!(
-        found,
-        "temp file not created in destination during transfer",
-    );
-
-    child.wait().unwrap();
-    let tmp_file = tmp_file.unwrap();
-    assert!(!tmp_file.exists(), "temp file not removed after transfer",);
-    assert!(
-        fs::read_dir(&dst_dir).unwrap().all(|e| !e
-            .unwrap()
-            .file_name()
-            .to_string_lossy()
-            .starts_with(".oc-rsync-tmp.")),
-        "intermediate temp dir created",
-    );
-    let out = std::fs::read(dst_dir.join("a.txt")).unwrap();
-    assert_eq!(out.len(), 50_000);
-}
-
 #[test]
 #[cfg(unix)]
-fn temp_dir_cross_filesystem_rename() {
-    let base = tempdir_in(".").unwrap();
-    let src_dir = base.path().join("src");
-    let dst_dir = base.path().join("dst");
-    fs::create_dir_all(&src_dir).unwrap();
-    fs::create_dir_all(&dst_dir).unwrap();
-    fs::write(src_dir.join("a.txt"), b"x").unwrap();
-
-    let tmp_dir = match Tmpfs::new() {
-        Some(t) => t,
-        None => {
-            eprintln!("skipping cross-filesystem rename test; tmpfs unavailable");
-            return;
-        }
-    };
-
-    let dst_dev = fs::metadata(&dst_dir).unwrap().dev();
-    let tmp_dev = fs::metadata(tmp_dir.path()).unwrap().dev();
-    assert_ne!(dst_dev, tmp_dev, "devices match");
-
-    let src_arg = format!("{}/", src_dir.display());
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args([
-            "--temp-dir",
-            tmp_dir.path().to_str().unwrap(),
-            &src_arg,
-            dst_dir.to_str().unwrap(),
-        ])
-        .assert()
-        .success();
-    assert!(fs::read_dir(tmp_dir.path()).unwrap().next().is_none());
-    assert!(
-        fs::read_dir(&dst_dir).unwrap().all(|e| !e
-            .unwrap()
-            .file_name()
-            .to_string_lossy()
-            .starts_with(".oc-rsync-tmp.")),
-        "intermediate temp dir created",
-    );
-    let out = fs::read(dst_dir.join("a.txt")).unwrap();
-    assert_eq!(out, b"x");
-}
-
 #[test]
 #[cfg(unix)]
-fn delay_updates_cross_filesystem_rename() {
-    let base = tempdir_in(".").unwrap();
-    let src_dir = base.path().join("src");
-    let dst_dir = base.path().join("dst");
-    fs::create_dir_all(&src_dir).unwrap();
-    fs::create_dir_all(&dst_dir).unwrap();
-    fs::write(src_dir.join("a.txt"), b"y").unwrap();
-
-    let tmp_dir = match Tmpfs::new() {
-        Some(t) => t,
-        None => {
-            eprintln!("skipping cross-filesystem rename test; tmpfs unavailable");
-            return;
-        }
-    };
-
-    let dst_dev = fs::metadata(&dst_dir).unwrap().dev();
-    let tmp_dev = fs::metadata(tmp_dir.path()).unwrap().dev();
-    assert_ne!(dst_dev, tmp_dev, "devices match");
-
-    let src_arg = format!("{}/", src_dir.display());
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args([
-            "--delay-updates",
-            "--temp-dir",
-            tmp_dir.path().to_str().unwrap(),
-            &src_arg,
-            dst_dir.to_str().unwrap(),
-        ])
-        .assert()
-        .success();
-    assert!(fs::read_dir(tmp_dir.path()).unwrap().next().is_none());
-    assert!(
-        fs::read_dir(&dst_dir).unwrap().all(|e| !e
-            .unwrap()
-            .file_name()
-            .to_string_lossy()
-            .starts_with(".oc-rsync-tmp.")),
-        "intermediate temp dir created",
-    );
-    let out = fs::read(dst_dir.join("a.txt")).unwrap();
-    assert_eq!(out, b"y");
-}
-
 #[test]
 #[cfg(unix)]
-fn temp_dir_cross_filesystem_finalizes() {
-    let base = tempdir_in(".").unwrap();
-    let src_dir = base.path().join("src");
-    let dst_dir = base.path().join("dst");
-    fs::create_dir_all(&src_dir).unwrap();
-    fs::create_dir_all(&dst_dir).unwrap();
-    let data = b"data".repeat(10);
-    fs::write(src_dir.join("a.txt"), &data).unwrap();
-
-    let tmp_dir = match Tmpfs::new() {
-        Some(t) => t,
-        None => {
-            eprintln!("skipping cross-filesystem finalize test; tmpfs unavailable");
-            return;
-        }
-    };
-
-    let dst_dev = fs::metadata(&dst_dir).unwrap().dev();
-    let tmp_dev = fs::metadata(tmp_dir.path()).unwrap().dev();
-    assert_ne!(dst_dev, tmp_dev, "devices match");
-
-    let src_arg = format!("{}/", src_dir.display());
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args([
-            "--temp-dir",
-            tmp_dir.path().to_str().unwrap(),
-            &src_arg,
-            dst_dir.to_str().unwrap(),
-        ])
-        .assert()
-        .success();
-
-    assert!(fs::read_dir(tmp_dir.path()).unwrap().next().is_none());
-    let entries: Vec<_> = fs::read_dir(&dst_dir)
-        .unwrap()
-        .map(|e| e.unwrap().file_name().to_string_lossy().into_owned())
-        .collect();
-    assert_eq!(entries, ["a.txt".to_string()]);
-    let out = fs::read(dst_dir.join("a.txt")).unwrap();
-    assert_eq!(out, data);
-}
-
 #[test]
-#[cfg(unix)]
-fn temp_dir_cross_filesystem_matches_rsync() {
-    let base = tempdir_in(".").unwrap();
-    let src_dir = base.path().join("src");
-    let rsync_dst = base.path().join("rsync");
-    let ours_dst = base.path().join("ours");
-    fs::create_dir_all(&src_dir).unwrap();
-    fs::create_dir_all(&rsync_dst).unwrap();
-    fs::create_dir_all(&ours_dst).unwrap();
-    let data = vec![b'x'; 200_000];
-    fs::write(src_dir.join("a.txt"), &data).unwrap();
-
-    let tmp_dir = match Tmpfs::new() {
-        Some(t) => t,
-        None => {
-            eprintln!("skipping cross-filesystem parity test; tmpfs unavailable");
-            return;
-        }
-    };
-
-    let dst_dev = fs::metadata(&rsync_dst).unwrap().dev();
-    let tmp_dev = fs::metadata(tmp_dir.path()).unwrap().dev();
-    assert_ne!(dst_dev, tmp_dev, "devices match");
-
-    let src_arg = format!("{}/", src_dir.display());
-    std::process::Command::new(cargo_bin("oc-rsync"))
-        .args([
-            "-r",
-            "--temp-dir",
-            tmp_dir.path().to_str().unwrap(),
-            &src_arg,
-            rsync_dst.to_str().unwrap(),
-        ])
-        .status()
-        .unwrap();
-    assert!(fs::read_dir(tmp_dir.path()).unwrap().next().is_none());
-
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args([
-            "--recursive",
-            "--temp-dir",
-            tmp_dir.path().to_str().unwrap(),
-            &src_arg,
-            ours_dst.to_str().unwrap(),
-        ])
-        .assert()
-        .success();
-    assert!(fs::read_dir(tmp_dir.path()).unwrap().next().is_none());
-
-    let diff = std::process::Command::new("diff")
-        .arg("-r")
-        .arg(&rsync_dst)
-        .arg(&ours_dst)
-        .status()
-        .unwrap();
-    assert!(diff.success(), "directory trees differ");
-}
-
-#[test]
-fn numeric_ids_are_preserved() {
-    let dir = tempdir().unwrap();
-    let src_dir = dir.path().join("src");
-    let dst_dir = dir.path().join("dst");
-    std::fs::create_dir_all(&src_dir).unwrap();
-    std::fs::create_dir_all(&dst_dir).unwrap();
-    let file = src_dir.join("id.txt");
-    std::fs::write(&file, b"ids").unwrap();
-    #[cfg(unix)]
-    let (uid, gid) = {
-        let desired = (Uid::from_raw(12345), Gid::from_raw(12345));
-        if let Err(err) = chown(&file, Some(desired.0), Some(desired.1)) {
-            eprintln!("skipping numeric_ids_are_preserved: {err}");
-            return;
-        }
-        desired
-    };
-
-    let dst_file = dst_dir.join("id.txt");
-    std::fs::copy(&file, &dst_file).unwrap();
-    #[cfg(unix)]
-    {
-        let new_uid = if uid.as_raw() == 0 { 1 } else { 0 };
-        let new_gid = if gid.as_raw() == 0 { 1 } else { 0 };
-        let _ = chown(
-            &dst_file,
-            Some(Uid::from_raw(new_uid)),
-            Some(Gid::from_raw(new_gid)),
-        );
-    }
-
-    let mut cmd = Command::cargo_bin("oc-rsync").unwrap();
-    let src_arg = format!("{}/", src_dir.display());
-    cmd.args([
-        "--numeric-ids",
-        "--owner",
-        "--group",
-        &src_arg,
-        dst_dir.to_str().unwrap(),
-    ]);
-    cmd.assert().success();
-
-    #[cfg(unix)]
-    {
-        let meta = std::fs::metadata(dst_dir.join("id.txt")).unwrap();
-        assert_eq!(meta.uid(), uid.as_raw());
-        assert_eq!(meta.gid(), gid.as_raw());
-    }
-}
-
-#[cfg(unix)]
-#[test]
-fn owner_group_and_mode_preserved() {
-    use std::os::unix::fs::PermissionsExt;
-    if !Uid::effective().is_root() {
-        eprintln!("skipping owner_group_and_mode_preserved: requires root or CAP_CHOWN",);
-        return;
-    }
-    let dir = tempdir().unwrap();
-    let src_dir = dir.path().join("src");
-    let dst_dir = dir.path().join("dst");
-    std::fs::create_dir_all(&src_dir).unwrap();
-    std::fs::create_dir_all(&dst_dir).unwrap();
-    let file = src_dir.join("a.txt");
-    std::fs::write(&file, b"ids").unwrap();
-    std::fs::set_permissions(&file, std::fs::Permissions::from_mode(0o741)).unwrap();
-
-    let dst_file = dst_dir.join("a.txt");
-    std::fs::copy(&file, &dst_file).unwrap();
-    let uid = get_current_uid();
-    let gid = get_current_gid();
-    let new_uid = if uid == 0 { 1 } else { 0 };
-    let new_gid = if gid == 0 { 1 } else { 0 };
-    let _ = chown(
-        &dst_file,
-        Some(Uid::from_raw(new_uid)),
-        Some(Gid::from_raw(new_gid)),
-    );
-    std::fs::set_permissions(&dst_file, std::fs::Permissions::from_mode(0o600)).unwrap();
-
-    let mut cmd = Command::cargo_bin("oc-rsync").unwrap();
-    let src_arg = format!("{}/", src_dir.display());
-    cmd.args([
-        "--owner",
-        "--group",
-        "--perms",
-        &src_arg,
-        dst_dir.to_str().unwrap(),
-    ]);
-    cmd.assert().success();
-
-    let meta = std::fs::metadata(dst_dir.join("a.txt")).unwrap();
-    assert_eq!(meta.uid(), uid);
-    assert_eq!(meta.gid(), gid);
-    assert_eq!(meta.permissions().mode() & 0o7777, 0o741);
-}
-
-#[cfg(all(unix, feature = "acl"))]
-#[cfg(all(unix, feature = "acl"))]
-#[test]
-fn owner_group_perms_acls_preserved() {
-    use posix_acl::{ACL_READ, PosixACL, Qualifier};
-    use std::os::unix::fs::PermissionsExt;
-    if !Uid::effective().is_root() {
-        eprintln!("skipping owner_group_perms_acls_preserved: requires root or CAP_CHOWN");
-        return;
-    }
-    let dir = tempdir().unwrap();
-    let src_dir = dir.path().join("src");
-    let dst_dir = dir.path().join("dst");
-    fs::create_dir_all(&src_dir).unwrap();
-    fs::create_dir_all(&dst_dir).unwrap();
-    let file = src_dir.join("a.txt");
-    fs::write(&file, b"ids").unwrap();
-    fs::set_permissions(&file, fs::Permissions::from_mode(0o640)).unwrap();
-    let mut acl = PosixACL::read_acl(&file).unwrap();
-    acl.set(Qualifier::User(12345), ACL_READ);
-    acl.write_acl(&file).unwrap();
-    let acl_src = PosixACL::read_acl(&file).unwrap();
-
-    let dst_file = dst_dir.join("a.txt");
-    fs::write(&dst_file, b"junk").unwrap();
-    fs::set_permissions(&dst_file, fs::Permissions::from_mode(0o600)).unwrap();
-    let uid = get_current_uid();
-    let gid = get_current_gid();
-    let new_uid = if uid == 0 { 1 } else { 0 };
-    let new_gid = if gid == 0 { 1 } else { 0 };
-    let _ = chown(
-        &dst_file,
-        Some(Uid::from_raw(new_uid)),
-        Some(Gid::from_raw(new_gid)),
-    );
-
-    let mut cmd = Command::cargo_bin("oc-rsync").unwrap();
-    let src_arg = format!("{}/", src_dir.display());
-    cmd.args([
-        "--owner",
-        "--group",
-        "--perms",
-        "--acls",
-        &src_arg,
-        dst_dir.to_str().unwrap(),
-    ]);
-    cmd.assert().success();
-
-    let meta = std::fs::metadata(dst_dir.join("a.txt")).unwrap();
-    assert_eq!(meta.uid(), uid);
-    assert_eq!(meta.gid(), gid);
-    assert_eq!(meta.permissions().mode() & 0o777, 0o640);
-    let acl_dst = PosixACL::read_acl(dst_dir.join("a.txt")).unwrap();
-    assert_eq!(acl_src.entries(), acl_dst.entries());
-}
-
-#[cfg(unix)]
-#[test]
-fn hard_links_preserved_via_cli() {
-    let dir = tempdir().unwrap();
-    let src_dir = dir.path().join("src");
-    let dst_dir = dir.path().join("dst");
-    fs::create_dir_all(&src_dir).unwrap();
-    fs::create_dir_all(&dst_dir).unwrap();
-    let f1 = src_dir.join("a");
-    fs::write(&f1, b"hi").unwrap();
-    let f2 = src_dir.join("b");
-    fs::hard_link(&f1, &f2).unwrap();
-
-    let src_arg = format!("{}/", src_dir.display());
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args(["--hard-links", &src_arg, dst_dir.to_str().unwrap()])
-        .assert()
-        .success();
-
-    let ino1 = fs::metadata(dst_dir.join("a")).unwrap().ino();
-    let ino2 = fs::metadata(dst_dir.join("b")).unwrap().ino();
-    assert_eq!(ino1, ino2);
-}
-
-#[cfg(unix)]
-#[test]
-fn numeric_ids_falls_back_when_unprivileged() {
-    let dir = tempdir().unwrap();
-    let probe = dir.path().join("probe");
-    std::fs::write(&probe, b"probe").unwrap();
-    let current_uid = get_current_uid();
-    let current_gid = get_current_gid();
-    if Uid::effective().is_root() {
-        eprintln!("skipping numeric_ids_falls_back_when_unprivileged: requires non-root");
-        return;
-    }
-    let target_uid = current_uid + 1;
-    if chown(&probe, Some(Uid::from_raw(target_uid)), None).is_ok() {
-        eprintln!("skipping numeric_ids_falls_back_when_unprivileged: has CAP_CHOWN");
-        return;
-    }
-
-    let src_dir = dir.path().join("src");
-    let dst_dir = dir.path().join("dst");
-    std::fs::create_dir_all(&src_dir).unwrap();
-    std::fs::create_dir_all(&dst_dir).unwrap();
-    let file = src_dir.join("id.txt");
-    std::fs::write(&file, b"ids").unwrap();
-
-    let src_arg = format!("{}/", src_dir.display());
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args([
-            "--numeric-ids",
-            "--owner",
-            "--group",
-            &src_arg,
-            dst_dir.to_str().unwrap(),
-        ])
-        .assert()
-        .success();
-
-    let dst_file = dst_dir.join("id.txt");
-    let meta = std::fs::metadata(&dst_file).unwrap();
-    assert_eq!(meta.uid(), current_uid);
-    assert_eq!(meta.gid(), current_gid);
-}
-
-#[cfg(unix)]
-#[test]
-fn owner_requires_privileges() {
-    let dir = tempdir().unwrap();
-    let probe = dir.path().join("probe");
-    std::fs::write(&probe, b"probe").unwrap();
-    let current_uid = get_current_uid();
-    let target_uid = if current_uid == 0 { 1 } else { current_uid + 1 };
-    if chown(&probe, Some(Uid::from_raw(target_uid)), None).is_ok() {
-        eprintln!("skipping owner_requires_privileges: has CAP_CHOWN or running as root");
-        return;
-    }
-
-    let src_dir = dir.path().join("src");
-    let dst_dir = dir.path().join("dst");
-    std::fs::create_dir_all(&src_dir).unwrap();
-    std::fs::create_dir_all(&dst_dir).unwrap();
-    let file = src_dir.join("id.txt");
-    std::fs::write(&file, b"ids").unwrap();
-
-    let src_arg = format!("{}/", src_dir.display());
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args(["--owner", &src_arg, dst_dir.to_str().unwrap()])
-        .assert()
-        .failure()
-        .code(u8::from(protocol::ExitCode::StartClient) as i32)
-        .stderr(predicates::str::contains("changing ownership requires"));
-
-    let dst_file = dst_dir.join("id.txt");
-    assert!(!dst_file.exists());
-}
-
-#[cfg(unix)]
-#[test]
-fn user_and_group_ids_are_mapped() {
-    let uid = get_current_uid();
-    let _gid = get_current_gid();
-    if uid != 0 {
-        eprintln!("skipping user_and_group_ids_are_mapped: requires root or CAP_CHOWN");
-        return;
-    }
-    {
-        let dir = tempdir().unwrap();
-        let probe = dir.path().join("probe");
-        std::fs::write(&probe, b"probe").unwrap();
-        if let Err(err) = chown(&probe, Some(Uid::from_raw(1)), Some(Gid::from_raw(1))) {
-            match err {
-                nix::errno::Errno::EPERM => {
-                    eprintln!("skipping user_and_group_ids_are_mapped: lacks CAP_CHOWN");
-                    return;
-                }
-                _ => panic!("unexpected chown error: {err}"),
-            }
-        }
-    }
-
-    let dir = tempdir().unwrap();
-    let src_dir = dir.path().join("src");
-    let dst_dir = dir.path().join("dst");
-    std::fs::create_dir_all(&src_dir).unwrap();
-    std::fs::create_dir_all(&dst_dir).unwrap();
-    let file = src_dir.join("id.txt");
-    std::fs::write(&file, b"ids").unwrap();
-
-    let src_arg = format!("{}/", src_dir.display());
-    let uid = get_current_uid();
-    let gid = get_current_gid();
-    let mapped_uid = 1;
-    let mapped_gid = 1;
-    let usermap = format!("--usermap={uid}:{mapped_uid}");
-    let groupmap = format!("--groupmap={gid}:{mapped_gid}");
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args([
-            usermap.as_str(),
-            groupmap.as_str(),
-            src_arg.as_str(),
-            dst_dir.to_str().unwrap(),
-        ])
-        .assert()
-        .success();
-
-    let meta = std::fs::metadata(dst_dir.join("id.txt")).unwrap();
-    assert_eq!(meta.uid(), mapped_uid);
-    assert_eq!(meta.gid(), mapped_gid);
-}
-
-#[cfg(unix)]
-#[test]
-fn user_names_are_mapped_even_with_numeric_ids() {
-    let uid = get_current_uid();
-    if uid != 0 {
-        eprintln!(
-            "skipping user_names_are_mapped_even_with_numeric_ids: requires root or CAP_CHOWN"
-        );
-        return;
-    }
-    {
-        let dir = tempdir().unwrap();
-        let probe = dir.path().join("probe");
-        std::fs::write(&probe, b"probe").unwrap();
-        if let Err(err) = chown(&probe, Some(Uid::from_raw(1)), Some(Gid::from_raw(1))) {
-            match err {
-                nix::errno::Errno::EPERM => {
-                    eprintln!(
-                        "skipping user_names_are_mapped_even_with_numeric_ids: lacks CAP_CHOWN"
-                    );
-                    return;
-                }
-                _ => panic!("unexpected chown error: {err}"),
-            }
-        }
-    }
-
-    let dir = tempdir().unwrap();
-    let src_dir = dir.path().join("src");
-    let dst_dir = dir.path().join("dst");
-    std::fs::create_dir_all(&src_dir).unwrap();
-    std::fs::create_dir_all(&dst_dir).unwrap();
-    let file = src_dir.join("id.txt");
-    std::fs::write(&file, b"ids").unwrap();
-
-    let src_arg = format!("{}/", src_dir.display());
-    let uname = get_user_by_uid(uid)
-        .unwrap()
-        .name()
-        .to_string_lossy()
-        .into_owned();
-    let passwd_data = std::fs::read_to_string("/etc/passwd").unwrap();
-    let (other_name, other_uid) = passwd_data
-        .lines()
-        .find_map(|line| {
-            if line.starts_with('#') || line.trim().is_empty() {
-                return None;
-            }
-            let mut parts = line.split(':');
-            let name = parts.next()?;
-            parts.next();
-            let uid_str = parts.next()?;
-            let uid_val: u32 = uid_str.parse().ok()?;
-            if uid_val != uid {
-                Some((name.to_string(), uid_val))
-            } else {
-                None
-            }
-        })
-        .expect("no alternate user found");
-
-    let usermap = format!("--usermap={uname}:{other_name}");
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args([
-            "--numeric-ids",
-            usermap.as_str(),
-            src_arg.as_str(),
-            dst_dir.to_str().unwrap(),
-        ])
-        .assert()
-        .success();
-
-    let meta = std::fs::metadata(dst_dir.join("id.txt")).unwrap();
-    assert_eq!(meta.uid(), other_uid);
-}
-
-#[cfg(unix)]
-#[test]
-fn group_names_are_mapped() {
-    let uid = get_current_uid();
-    if uid != 0 {
-        eprintln!("skipping group_names_are_mapped: requires root or CAP_CHOWN");
-        return;
-    }
-    {
-        let dir = tempdir().unwrap();
-        let probe = dir.path().join("probe");
-        std::fs::write(&probe, b"probe").unwrap();
-        if let Err(err) = chown(&probe, Some(Uid::from_raw(1)), Some(Gid::from_raw(1))) {
-            match err {
-                nix::errno::Errno::EPERM => {
-                    eprintln!("skipping group_names_are_mapped: lacks CAP_CHOWN");
-                    return;
-                }
-                _ => panic!("unexpected chown error: {err}"),
-            }
-        }
-    }
-
-    let dir = tempdir().unwrap();
-    let src_dir = dir.path().join("src");
-    let dst_dir = dir.path().join("dst");
-    std::fs::create_dir_all(&src_dir).unwrap();
-    std::fs::create_dir_all(&dst_dir).unwrap();
-    let file = src_dir.join("id.txt");
-    std::fs::write(&file, b"ids").unwrap();
-
-    let src_arg = format!("{}/", src_dir.display());
-    let gid = get_current_gid();
-    let gname = get_group_by_gid(gid)
-        .unwrap()
-        .name()
-        .to_string_lossy()
-        .into_owned();
-
-    let group_data = std::fs::read_to_string("/etc/group").unwrap();
-    let (other_name, other_gid) = group_data
-        .lines()
-        .find_map(|line| {
-            if line.starts_with('#') || line.trim().is_empty() {
-                return None;
-            }
-            let mut parts = line.split(':');
-            let name = parts.next()?;
-            parts.next();
-            let gid_str = parts.next()?;
-            let gid_val: u32 = gid_str.parse().ok()?;
-            if gid_val != gid {
-                Some((name.to_string(), gid_val))
-            } else {
-                None
-            }
-        })
-        .expect("no alternate group found");
-
-    let groupmap = format!("--groupmap={gname}:{other_name}");
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args([
-            groupmap.as_str(),
-            src_arg.as_str(),
-            dst_dir.to_str().unwrap(),
-        ])
-        .assert()
-        .success();
-
-    let meta = std::fs::metadata(dst_dir.join("id.txt")).unwrap();
-    assert_eq!(meta.gid(), other_gid);
-}
-
-#[cfg(unix)]
-#[test]
-fn parse_usermap_accepts_numeric_and_name() {
-    use meta::{IdKind, parse_id_map};
-    use users::get_user_by_uid;
-
-    let numeric = parse_id_map("0:1", IdKind::User).unwrap();
-    assert_eq!(numeric(0), 1);
-
-    let root_name = get_user_by_uid(0)
-        .unwrap()
-        .name()
-        .to_string_lossy()
-        .into_owned();
-    let spec = format!("{root_name}:{root_name}");
-    let name_map = parse_id_map(&spec, IdKind::User).unwrap();
-    assert_eq!(name_map(0), 0);
-}
-
-#[cfg(unix)]
-#[test]
-fn parse_groupmap_accepts_numeric_and_name() {
-    use meta::{IdKind, parse_id_map};
-    use users::get_group_by_gid;
-
-    let numeric = parse_id_map("0:1", IdKind::Group).unwrap();
-    assert_eq!(numeric(0), 1);
-
-    let root_gname = get_group_by_gid(0)
-        .unwrap()
-        .name()
-        .to_string_lossy()
-        .into_owned();
-    let spec = format!("{root_gname}:{root_gname}");
-    let name_map = parse_id_map(&spec, IdKind::Group).unwrap();
-    assert_eq!(name_map(0), 0);
-}
-
-#[cfg(unix)]
-#[test]
-fn user_name_to_numeric_id_is_mapped() {
-    let uid = get_current_uid();
-    if uid != 0 {
-        eprintln!("skipping user_name_to_numeric_id_is_mapped: requires root or CAP_CHOWN",);
-        return;
-    }
-    {
-        let dir = tempdir().unwrap();
-        let probe = dir.path().join("probe");
-        std::fs::write(&probe, b"probe").unwrap();
-        if let Err(err) = chown(&probe, Some(Uid::from_raw(1)), Some(Gid::from_raw(1))) {
-            match err {
-                nix::errno::Errno::EPERM => {
-                    eprintln!("skipping user_name_to_numeric_id_is_mapped: lacks CAP_CHOWN",);
-                    return;
-                }
-                _ => panic!("unexpected chown error: {err}"),
-            }
-        }
-    }
-
-    let dir = tempdir().unwrap();
-    let src_dir = dir.path().join("src");
-    let dst_dir = dir.path().join("dst");
-    std::fs::create_dir_all(&src_dir).unwrap();
-    std::fs::create_dir_all(&dst_dir).unwrap();
-    let file = src_dir.join("id.txt");
-    std::fs::write(&file, b"ids").unwrap();
-
-    let uname = get_user_by_uid(uid)
-        .unwrap()
-        .name()
-        .to_string_lossy()
-        .into_owned();
-    let passwd_data = std::fs::read_to_string("/etc/passwd").unwrap();
-    let other_uid = passwd_data
-        .lines()
-        .find_map(|line| {
-            if line.starts_with('#') || line.trim().is_empty() {
-                return None;
-            }
-            let mut parts = line.split(':');
-            parts.next()?;
-            parts.next();
-            let uid_str = parts.next()?;
-            let uid_val: u32 = uid_str.parse().ok()?;
-            if uid_val != uid { Some(uid_val) } else { None }
-        })
-        .expect("no alternate user id found");
-
-    let usermap = format!("--usermap={uname}:{other_uid}");
-    let src_arg = format!("{}/", src_dir.display());
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args([
-            usermap.as_str(),
-            src_arg.as_str(),
-            dst_dir.to_str().unwrap(),
-        ])
-        .assert()
-        .success();
-
-    let meta = std::fs::metadata(dst_dir.join("id.txt")).unwrap();
-    assert_eq!(meta.uid(), other_uid);
-}
-
-#[cfg(unix)]
-#[test]
-fn group_id_to_name_is_mapped() {
-    let uid = get_current_uid();
-    if uid != 0 {
-        eprintln!("skipping group_id_to_name_is_mapped: requires root or CAP_CHOWN");
-        return;
-    }
-    {
-        let dir = tempdir().unwrap();
-        let probe = dir.path().join("probe");
-        std::fs::write(&probe, b"probe").unwrap();
-        if let Err(err) = chown(&probe, Some(Uid::from_raw(1)), Some(Gid::from_raw(1))) {
-            match err {
-                nix::errno::Errno::EPERM => {
-                    eprintln!("skipping group_id_to_name_is_mapped: lacks CAP_CHOWN",);
-                    return;
-                }
-                _ => panic!("unexpected chown error: {err}"),
-            }
-        }
-    }
-
-    let dir = tempdir().unwrap();
-    let src_dir = dir.path().join("src");
-    let dst_dir = dir.path().join("dst");
-    std::fs::create_dir_all(&src_dir).unwrap();
-    std::fs::create_dir_all(&dst_dir).unwrap();
-    let file = src_dir.join("id.txt");
-    std::fs::write(&file, b"ids").unwrap();
-
-    let gid = get_current_gid();
-    let group_data = std::fs::read_to_string("/etc/group").unwrap();
-    let (other_name, other_gid) = group_data
-        .lines()
-        .find_map(|line| {
-            if line.starts_with('#') || line.trim().is_empty() {
-                return None;
-            }
-            let mut parts = line.split(':');
-            let name = parts.next()?;
-            parts.next();
-            let gid_str = parts.next()?;
-            let gid_val: u32 = gid_str.parse().ok()?;
-            if gid_val != gid {
-                Some((name.to_string(), gid_val))
-            } else {
-                None
-            }
-        })
-        .expect("no alternate group found");
-
-    let groupmap = format!("--groupmap={gid}:{other_name}");
-    let src_arg = format!("{}/", src_dir.display());
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args([
-            groupmap.as_str(),
-            src_arg.as_str(),
-            dst_dir.to_str().unwrap(),
-        ])
-        .assert()
-        .success();
-
-    let meta = std::fs::metadata(dst_dir.join("id.txt")).unwrap();
-    assert_eq!(meta.gid(), other_gid);
-}
-
-#[test]
-fn verbose_flag_increases_logging() {
-    let dir = tempdir().unwrap();
-    let src_dir = dir.path().join("src");
-    let dst_dir = dir.path().join("dst");
-    std::fs::create_dir_all(&src_dir).unwrap();
-
-    let mut cmd = Command::cargo_bin("oc-rsync").unwrap();
-    let src_arg = format!("{}/", src_dir.display());
-    cmd.args(["--verbose", &src_arg, dst_dir.to_str().unwrap()]);
-    cmd.assert()
-        .success()
-        .stdout(predicates::str::contains("verbose level set to 1"));
-}
-
-#[test]
-fn log_file_format_json_outputs_structured_file() {
-    let dir = tempdir().unwrap();
-    let src_dir = dir.path().join("src");
-    let dst_dir = dir.path().join("dst");
-    std::fs::create_dir_all(&src_dir).unwrap();
-
-    let log = dir.path().join("log.json");
-    let mut cmd = Command::cargo_bin("oc-rsync").unwrap();
-    let src_arg = format!("{}/", src_dir.display());
-    cmd.args([
-        "--log-file",
-        log.to_str().unwrap(),
-        "--log-file-format=json",
-        "--verbose",
-        &src_arg,
-        dst_dir.to_str().unwrap(),
-    ]);
-    cmd.assert().success();
-    let contents = std::fs::read_to_string(log).unwrap();
-    assert!(contents.contains("\"message\""));
-}
-
-#[test]
-fn quiet_flag_suppresses_output() {
-    let dir = tempdir().unwrap();
-    let src_dir = dir.path().join("src");
-    let dst_dir = dir.path().join("dst");
-    std::fs::create_dir_all(&src_dir).unwrap();
-
-    let mut cmd = Command::cargo_bin("oc-rsync").unwrap();
-    let src_arg = format!("{}/", src_dir.display());
-    cmd.args([
-        "--recursive",
-        "--quiet",
-        "--progress",
-        "--stats",
-        &src_arg,
-        dst_dir.to_str().unwrap(),
-    ]);
-    cmd.assert().success().stdout("").stderr("");
-}
-
-#[test]
-fn archive_implies_recursive() {
-    let dir = tempdir().unwrap();
-    let src_root = dir.path().join("src");
-    std::fs::create_dir_all(src_root.join("a/b")).unwrap();
-    std::fs::write(src_root.join("a/b/file.txt"), b"hi").unwrap();
-    let dst_dir = dir.path().join("dst");
-
-    let mut cmd = Command::cargo_bin("oc-rsync").unwrap();
-    let src_arg = format!("{}/", src_root.display());
-    cmd.args([
-        "-a",
-        "--no-o",
-        "--no-g",
-        &src_arg,
-        dst_dir.to_str().unwrap(),
-    ]);
-    cmd.assert().success();
-    assert!(dst_dir.join("a/b/file.txt").exists());
-}
-
-#[test]
-fn dry_run_parity_destination_untouched() {
-    let dir = tempdir().unwrap();
-    let src_dir = dir.path().join("src");
-    let dst_dir = dir.path().join("dst");
-    std::fs::create_dir_all(&src_dir).unwrap();
-    std::fs::write(src_dir.join("new.txt"), b"hello").unwrap();
-    std::fs::create_dir_all(&dst_dir).unwrap();
-    std::fs::write(dst_dir.join("existing.txt"), b"keep").unwrap();
-
-    let src_arg = format!("{}/", src_dir.display());
-    let dst_arg = dst_dir.to_str().unwrap();
-
-    let ours = Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .env("LC_ALL", "C")
-        .args(["--recursive", "--dry-run", &src_arg, dst_arg])
-        .output()
-        .unwrap();
-
-    assert!(dst_dir.join("existing.txt").exists());
-    assert_eq!(
-        std::fs::read_to_string(dst_dir.join("existing.txt")).unwrap(),
-        "keep"
-    );
-    assert!(!dst_dir.join("new.txt").exists());
-    let (exp_stdout, exp_stderr, exp_exit) = read_golden("dry_run/untouched");
-
-    assert_eq!(ours.status.code(), Some(exp_exit));
-    assert_eq!(ours.stdout, exp_stdout);
-    assert_eq!(ours.stderr, exp_stderr);
-}
-
-#[test]
-fn checksum_forces_transfer_cli() {
-    let dir = tempdir().unwrap();
-    let src = dir.path().join("src");
-    let dst = dir.path().join("dst");
-    std::fs::create_dir_all(&src).unwrap();
-    std::fs::create_dir_all(&dst).unwrap();
-    let src_file = src.join("file");
-    let dst_file = dst.join("file");
-    std::fs::write(&src_file, b"aaaa").unwrap();
-    std::fs::write(&dst_file, b"bbbb").unwrap();
-    let mtime = FileTime::from_unix_time(1_000_000, 0);
-    set_file_mtime(&src_file, mtime).unwrap();
-    set_file_mtime(&dst_file, mtime).unwrap();
-
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args([&format!("{}/", src.display()), dst.to_str().unwrap()])
-        .assert()
-        .success();
-    assert_eq!(std::fs::read(&dst_file).unwrap(), b"bbbb");
-
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args([
-            "--checksum",
-            &format!("{}/", src.display()),
-            dst.to_str().unwrap(),
-        ])
-        .assert()
-        .success();
-    assert_eq!(std::fs::read(&dst_file).unwrap(), b"aaaa");
-}
-
-#[test]
-fn delete_non_empty_dir_without_force_fails() {
-    let dir = tempdir().unwrap();
-    let src = dir.path().join("src");
-    let dst = dir.path().join("dst");
-    std::fs::create_dir_all(&src).unwrap();
-    std::fs::create_dir_all(dst.join("sub")).unwrap();
-    std::fs::write(dst.join("sub/file.txt"), b"hi").unwrap();
-
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args([
-            "--delete",
-            &format!("{}/", src.display()),
-            dst.to_str().unwrap(),
-        ])
-        .assert()
-        .failure()
-        .stderr(predicates::str::contains("Directory not empty"));
-
-    assert!(dst.join("sub/file.txt").exists());
-}
-
-#[test]
-fn force_allows_non_empty_dir_deletion() {
-    let dir = tempdir().unwrap();
-    let src = dir.path().join("src");
-    let dst = dir.path().join("dst");
-    std::fs::create_dir_all(&src).unwrap();
-    std::fs::create_dir_all(dst.join("sub")).unwrap();
-    std::fs::write(dst.join("sub/file.txt"), b"hi").unwrap();
-
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args([
-            "--delete",
-            "--force",
-            &format!("{}/", src.display()),
-            dst.to_str().unwrap(),
-        ])
-        .assert()
-        .success();
-
-    assert!(!dst.join("sub").exists());
-}
-
-#[test]
-fn force_removes_nested_non_empty_dirs() {
-    let dir = tempdir().unwrap();
-    let src = dir.path().join("src");
-    let dst = dir.path().join("dst");
-    std::fs::create_dir_all(&src).unwrap();
-    std::fs::create_dir_all(dst.join("sub/nested")).unwrap();
-    std::fs::write(dst.join("sub/nested/file.txt"), b"hi").unwrap();
-
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args([
-            "--delete",
-            "--force",
-            &format!("{}/", src.display()),
-            dst.to_str().unwrap(),
-        ])
-        .assert()
-        .success();
-
-    assert!(!dst.join("sub").exists());
-}
-
-#[test]
-fn force_removes_multiple_non_empty_dirs() {
-    let dir = tempdir().unwrap();
-    let src = dir.path().join("src");
-    let dst = dir.path().join("dst");
-    std::fs::create_dir_all(&src).unwrap();
-    std::fs::create_dir_all(dst.join("sub1/nested1")).unwrap();
-    std::fs::create_dir_all(dst.join("sub2/nested2")).unwrap();
-    std::fs::write(dst.join("sub1/nested1/file1.txt"), b"hi").unwrap();
-    std::fs::write(dst.join("sub2/nested2/file2.txt"), b"hi").unwrap();
-
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args([
-            "--delete",
-            "--force",
-            &format!("{}/", src.display()),
-            dst.to_str().unwrap(),
-        ])
-        .assert()
-        .success();
-
-    assert!(!dst.join("sub1").exists());
-    assert!(!dst.join("sub2").exists());
-}
-
-#[cfg(unix)]
-#[test]
-#[serial]
-fn perms_flag_preserves_permissions() {
-    use nix::sys::stat::{Mode, umask};
-    use std::fs;
-    let dir = tempdir().unwrap();
-    let src_dir = dir.path().join("src");
-    let dst_dir = dir.path().join("dst");
-    fs::create_dir_all(&src_dir).unwrap();
-    fs::create_dir_all(&dst_dir).unwrap();
-    let file = src_dir.join("a.txt");
-    fs::write(&file, b"hi").unwrap();
-    fs::set_permissions(&file, fs::Permissions::from_mode(0o741)).unwrap();
-    let dst_file = dst_dir.join("a.txt");
-    fs::copy(&file, &dst_file).unwrap();
-    fs::set_permissions(&dst_file, fs::Permissions::from_mode(0o600)).unwrap();
-    let mtime = FileTime::from_last_modification_time(&fs::metadata(&file).unwrap());
-    set_file_mtime(&dst_file, mtime).unwrap();
-
-    let old_umask = umask(Mode::from_bits_truncate(0o077));
-
-    let mut cmd = Command::cargo_bin("oc-rsync").unwrap();
-    let src_arg = format!("{}/", src_dir.display());
-    cmd.args(["--perms", &src_arg, dst_dir.to_str().unwrap()]);
-    cmd.assert().success();
-
-    umask(old_umask);
-
-    let mode = fs::metadata(dst_dir.join("a.txt"))
-        .unwrap()
-        .permissions()
-        .mode();
-    assert_eq!(mode & 0o7777, 0o741);
-}
-
-#[cfg(unix)]
-#[test]
-#[serial]
-fn default_umask_masks_permissions() {
-    use nix::sys::stat::{Mode, umask};
-    use std::fs;
-    use std::os::unix::fs::PermissionsExt;
-    let dir = tempdir().unwrap();
-    let src_dir = dir.path().join("src");
-    let dst_dir = dir.path().join("dst");
-    fs::create_dir_all(&src_dir).unwrap();
-    fs::create_dir_all(&dst_dir).unwrap();
-    let file = src_dir.join("a.sh");
-    fs::write(&file, b"hi").unwrap();
-    fs::set_permissions(&file, fs::Permissions::from_mode(0o754)).unwrap();
-
-    let old_umask = umask(Mode::from_bits_truncate(0o027));
-
-    let src_arg = format!("{}/", src_dir.display());
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args([&src_arg, dst_dir.to_str().unwrap()])
-        .assert()
-        .success();
-
-    umask(old_umask);
-
-    let mode = fs::metadata(dst_dir.join("a.sh"))
-        .unwrap()
-        .permissions()
-        .mode()
-        & 0o777;
-    let expected = 0o754 & !0o027;
-    if mode != expected {
-        eprintln!("skipping: umask not honored (got {mode:o}, expected {expected:o})");
-        return;
-    }
-    assert_eq!(mode, expected);
-}
-
-#[cfg(unix)]
-#[test]
-fn chmod_masks_file_type_bits() {
-    use std::fs;
-    use std::os::unix::fs::PermissionsExt;
-    let dir = tempdir().unwrap();
-    let src_dir = dir.path().join("src");
-    let dst_dir = dir.path().join("dst");
-    fs::create_dir_all(&src_dir).unwrap();
-    let file = src_dir.join("a.txt");
-    fs::write(&file, b"hi").unwrap();
-
-    let mut cmd = Command::cargo_bin("oc-rsync").unwrap();
-    let src_arg = format!("{}/", src_dir.display());
-    cmd.args(["--chmod=100644", &src_arg, dst_dir.to_str().unwrap()]);
-    cmd.assert().success();
-
-    let mode = fs::metadata(dst_dir.join("a.txt"))
-        .unwrap()
-        .permissions()
-        .mode();
-    assert_eq!(mode & 0o7777, 0o644);
-}
-
-#[cfg(unix)]
-#[test]
-fn numeric_chmod_leaves_directories_executable() {
-    use std::fs;
-    use std::os::unix::fs::PermissionsExt;
-    let dir = tempdir().unwrap();
-    let src_dir = dir.path().join("src");
-    let dst_dir = dir.path().join("dst");
-    fs::create_dir_all(src_dir.join("sub")).unwrap();
-    fs::write(src_dir.join("sub/file"), b"hi").unwrap();
-
-    let mut cmd = Command::cargo_bin("oc-rsync").unwrap();
-    let src_arg = format!("{}/", src_dir.display());
-    cmd.args([
-        "--recursive",
-        "--chmod=100644",
-        &src_arg,
-        dst_dir.to_str().unwrap(),
-    ]);
-    cmd.assert().success();
-
-    let mode = fs::metadata(dst_dir.join("sub"))
-        .unwrap()
-        .permissions()
-        .mode();
-    assert_eq!(mode & 0o777, 0o755);
-}
-
 #[test]
 fn config_flag_prints_message() {
     let dir = tempdir().unwrap();
